Create an even more minimal run file; tweak init script

+++ src/qry/PBSPlugin.scala
@@ -22,7 +22,7 @@ object PBS {
       case MEMORY_REGEX(x, separator) => x
       case _ => memory
     }
-    "mem=" + mem + ",ppn=" + cores + ",nodes=1"
+    "mem=" + mem + ":nodes=1:" + "cores=" + cores
   }
 
   def run(bashCmd:String, execDir:Option[String]):Option[Int] = {
@@ -123,7 +123,7 @@ object PBS {
 
   /** The header of the script to run the PBS job with. The command goes between the header and footer. */
   private val header = """
-set -x
+# set -x
 
 USERNAME=`whoami`
 HOSTNAME=`hostname | sed -r -e 's/\.[^\.]+\.(edu|com)//i'`
@@ -161,7 +161,7 @@ while [ $sleep_counter -lt 20 ] && [ ! -d "$LOG_DIR" ]; do
 done
 
 cd "$WD"
-echo `date +"%a %b %d %k:%M:%S %Y"`: Started on ${HOSTNAME} in ${PBS_O_QUEUE} queue. >> "$LOG_DIR/_qsub.log"
+echo `date +"%a %b %d %k:%M:%S %Y"`: Started on ${HOSTNAME} in ${PBS_O_QUEUE} queue. >> "$LOG_DIR/_pbs.log"
 
 # actually run the command
 ( """
@@ -169,11 +169,13 @@ echo `date +"%a %b %d %k:%M:%S %Y"`: Started on ${HOSTNAME} in ${PBS_O_QUEUE} qu
   /** The footer of the script to run the PBS job with. The command goes between the header and footer. */
   private val footer = """ > "$LOG_DIR/_stdout.log" 2> "$LOG_DIR/_stderr.log" )
 
-echo `date +"%a %b %d %k:%M:%S %Y"`: Completed on ${HOSTNAME} >> '%(qsub_log)s'
+echo `date +"%a %b %d %k:%M:%S %Y"`: Completed on ${HOSTNAME} >> "$LOG_DIR/_pbs.log"
 
 # cleanup the temp files we made -- this should really be done in the epilog
 # if we ever figure out how to get PBS to run them
-rm -rf $TMP
+if [[ $TMP == /tmp/* ]]; then  # rm -rf is very dangerous; make sure this is a tmp dir
+  rm -rf $TMP
+fi
 """
 }
